implemented user repository
